Made both http and rtsp use the same format to report buffering progress using a 	single int. Rtsp was using 2 ints (one was an unused constant 1). 	The percentage is an int vs uint32 all the way to make it more portable. 	Used void instead of OsclAny to make the code more readable. 	Similarly size_t is used instead of uint32.
	* pvmi/pvmf/include/pvmf_event_handling.h:
	Replace manual copy loops with memcpy.

	* nodes/streaming/jitterbuffernode/src/pvmf_jitter_buffer_node.cpp:
	Uses an int to report the buffering percentage.
	Removed unused 1st constant (alsways set to 1) in the value buffer.

	* nodes/streaming/jitterbuffernode/jitterbuffer/common/include/pvmf_jitter_buffer.h:
	* nodes/pvprotocolenginenode/download_protocols/progressive_streaming/src/pvmf_protocol_engine_node_progressive_streaming.cpp:
	The buffering status is now an int (more portable).

	* android/playerdriver.cpp:
	Added missing oscl includes.
	Remove the logic that dealt with the 2 kind of buffering event (rtsp vs http).
	The new buffer status update is made of just the percent value (like http previously).
	The value is an int and not an unit32.

+++ nodes/pvprotocolenginenode/protocol_common/src/pvmf_protocol_engine_node_common.h
@@ -1184,7 +1184,9 @@ enum EventReporterSupportObjectType
 class EventReporterObserver
 {
     public:
-        virtual void ReportEvent(PVMFEventType aEventType, OsclAny* aEventData = NULL, const int32 aEventCode = 0, OsclAny* aEventLocalBuffer = NULL, const uint32 aEventLocalBufferSize = 0) = 0;
+    // TODO: Change the buffer to be const void*.
+    // TODO: Why do we have event data AND event buffer?
+        virtual void ReportEvent(PVMFEventType aEventType, OsclAny* aEventData = NULL, const int32 aEventCode = 0, OsclAny* aEventLocalBuffer = NULL, const size_t aEventLocalBufferSize = 0) = 0;
         virtual void NotifyContentTooLarge() = 0;
         virtual uint32 GetObserverState() = 0;
 };
@@ -1725,4 +1727,3 @@ class PVMFProtocolEngineNodeTimer
 };
 
 #endif
-
